+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 - [feature:add User-Agent:polaris for healthyCheck api.](https://github.com/Tencent/spring-cloud-tencent/pull/872)
 - [optimize:optimize ServiceRuleManager.](https://github.com/Tencent/spring-cloud-tencent/pull/877)
 - [refactor:refactor stat module.](https://github.com/Tencent/spring-cloud-tencent/pull/880)
+- [feature:add polaris circuit breaker support.](https://github.com/Tencent/spring-cloud-tencent/pull/882)